+++ src/Application/Command/PaymentFetchCommand.php
@@ -6,6 +6,8 @@ namespace App\Application\Command;
 
 use App\Application\DTO\PaymentFetchDTO;
 use App\Application\Exception\FetchingException;
+use App\Application\Exception\ReportNotFoundException;
+use App\Application\Exception\ReportNotCompleteException;
 use App\Domain\Model\PaymentReport;
 use App\Domain\Repository\PaymentReportRepository;
 use App\Domain\Repository\PaymentRepository;
@@ -13,20 +15,11 @@ use Psr\Log\LoggerInterface;
 
 final class PaymentFetchCommand
 {
-    private PaymentReportRepository $paymentReportRepository;
-
-    private PaymentRepository $paymentRepository;
-
-    private LoggerInterface $logger;
-
     public function __construct(
-        PaymentReportRepository $paymentReportRepository,
-        PaymentRepository $paymentRepository,
-        LoggerInterface $logger
+        private readonly PaymentReportRepository $paymentReportRepository,
+        private readonly PaymentRepository $paymentRepository,
+        private readonly LoggerInterface $logger
     ) {
-        $this->paymentReportRepository = $paymentReportRepository;
-        $this->paymentRepository = $paymentRepository;
-        $this->logger = $logger;
     }
 
     public function execute(int $timestamp, ?int $limit = null, ?int $offset = null): PaymentFetchDTO
@@ -35,8 +28,12 @@ final class PaymentFetchCommand
 
         $report = $this->paymentReportRepository->fetch(PaymentReport::timestampToId($timestamp));
 
+        if (null === $report) {
+            throw new ReportNotFoundException($timestamp);
+        }
+
         if (!$report->isComplete()) {
-            throw new FetchingException(sprintf('Report #%d is not complete yet.', $report->getId()));
+            throw new ReportNotCompleteException($timestamp);
         }
 
         if (!$report->isCalculated()) {